pyproject.toml
[project]
name = "workshopdjango2526-3ia2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.11.0",
    "django>=6.0",
    "fastmcp>=2.13.3",
]
